# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pickleball Hub API",
        "description": "API documentation for Pickleball Hub",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/time-slots": {
            "get": {
                "tags": [
                    "Time Slots"
                ],
                "summary": "Get all time slots",
                "description": "Returns all time slots along with their associated courts, sorted by latest first.",
                "operationId": "a15282e344e577a5fac2e8d43df9b4c8",
                "responses": {
                    "200": {
                        "description": "Time slots fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/owner/time-slots": {
            "post": {
                "tags": [
                    "Owner Time Slots"
                ],
                "summary": "Create a time slot",
                "description": "Allows an authenticated owner to create a time slot for their court.",
                "operationId": "4223f75141c57d0439f9fdb11fbf3246",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "court_id",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "court_id": {
                                        "description": "Court ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "description": "Time slot start time",
                                        "type": "string",
                                        "example": "06:00"
                                    },
                                    "end_time": {
                                        "description": "Time slot end time",
                                        "type": "string",
                                        "example": "07:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Time slot created successfully."
                    },
                    "400": {
                        "description": "Duplicate, overlapping, or inactive court."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Unauthorized or user is not an owner."
                    },
                    "404": {
                        "description": "Court not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/time-slots/{timeSlot}": {
            "get": {
                "tags": [
                    "Time Slots"
                ],
                "summary": "Get time slot details",
                "description": "Returns details of a specific time slot along with its associated court.",
                "operationId": "d9280c08dc1d41885041b9db42a962c0",
                "parameters": [
                    {
                        "name": "timeSlot",
                        "in": "path",
                        "description": "Time slot ID.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Time slot fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Time slot not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/owner/time-slots/{timeSlot}": {
            "put": {
                "tags": [
                    "Owner Time Slots"
                ],
                "summary": "Update a time slot",
                "description": "Allows an authenticated owner to update a time slot belonging to their court.",
                "operationId": "afc47221ccb7d34f99d588d53b8d39ca",
                "parameters": [
                    {
                        "name": "timeSlot",
                        "in": "path",
                        "description": "Time slot ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_time",
                                    "end_time",
                                    "status"
                                ],
                                "properties": {
                                    "start_time": {
                                        "description": "Updated start time",
                                        "type": "string",
                                        "example": "07:00"
                                    },
                                    "end_time": {
                                        "description": "Updated end time",
                                        "type": "string",
                                        "example": "08:00"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Time slot updated successfully."
                    },
                    "400": {
                        "description": "Time slot already exists."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Unauthorized or user is not an owner."
                    },
                    "404": {
                        "description": "Time slot not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Owner Time Slots"
                ],
                "summary": "Delete a time slot",
                "description": "Allows an authenticated owner to delete a time slot belonging to their court. A time slot cannot be deleted if it has bookings.",
                "operationId": "569fd01e9a54d04405622fb6a3cb0252",
                "parameters": [
                    {
                        "name": "timeSlot",
                        "in": "path",
                        "description": "Time slot ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Time slot deleted successfully."
                    },
                    "400": {
                        "description": "Cannot delete a booked time slot."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Unauthorized or user is not an owner."
                    },
                    "404": {
                        "description": "Time slot not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}